Added precondition to DeleteStudentClassroomAssignment methods

+++ ChildcareManagementStudio/DAL/StudentClassroomAssignmentDAL.cs
@@ -153,6 +153,11 @@ namespace ChildcareManagementStudio.DAL
         /// <param name="studentClassroomAssignment">The student/classroom assignment being deleted.</param>
         public void DeleteStudentClassroomAssignment(StudentClassroomAssignment studentClassroomAssignment)
         {
+            if (studentClassroomAssignment == null)
+            {
+                throw new ArgumentNullException("studentClassroomAssignment", "The StudentClassroomAssignment object cannot be null.");
+            }
+
             string deleteStatement =
                 "DELETE FROM StudentClassroomAssignment " +
                 "WHERE studentId = $studentId " +
